+++ include/tmc/spawn_task.hpp
@@ -38,7 +38,7 @@ class [[nodiscard("You must co_await aw_run_early. "
 template <typename Awaitable, typename Result> class aw_run_early_impl {
   std::coroutine_handle<> continuation;
   tmc::detail::type_erased_executor* continuation_executor;
-  std::atomic<int64_t> done_count;
+  std::atomic<ssize_t> done_count;
   tmc::detail::result_storage_t<Result> result;
 
   using AwaitableTraits = tmc::detail::get_awaitable_traits<Awaitable>;
@@ -114,7 +114,7 @@ public:
 template <typename Awaitable> class aw_run_early_impl<Awaitable, void> {
   std::coroutine_handle<> continuation;
   tmc::detail::type_erased_executor* continuation_executor;
-  std::atomic<int64_t> done_count;
+  std::atomic<ssize_t> done_count;
 
   using AwaitableTraits = tmc::detail::get_awaitable_traits<Awaitable>;
